# smarthire-backend/specs/001-smarthire-backend-platform/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SmartHire Backend API
  description: |
    Internal recruitment management platform API.
    All endpoints require a valid JWT bearer token except `/prescreen/**`, `/health`, `/docs`, and `/openapi.json`.
  version: "1.0.0"
  contact:
    name: SmartHire Platform Team
servers:
  - url: "http://localhost:8083"
    description: Local development
  - url: "https://smarthire.internal"
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseDto:
      type: object
      properties:
        response:
          type: array
          items: {}
        message:
          type: string
        exception:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
      required:
        - response
        - message
...
    ValidateSessionRequest:
      type: object
      required: [userName]
      properties:
        userName:
          type: string
          example: "jdoe@example.com"

    # -------------------------------------------------------------------------
    # Candidate
    # -------------------------------------------------------------------------
...
    ReportFilterRequest:
      type: object
      properties:
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        technologyId:
          type: integer
          format: int64
        buId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        interviewTypeId:
          type: integer
          format: int64

    # -------------------------------------------------------------------------
    # Prescreen
    # -------------------------------------------------------------------------
    # No request body — query params only

    # -------------------------------------------------------------------------
    # Teams Meeting
    # -------------------------------------------------------------------------
...
paths:
...
  /login/validateSession:
    post:
      tags: [Auth]
      summary: Validate Keycloak session and enforce single-session policy
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateSessionRequest"
      responses:
        "200":
          description: Session validated or duplicate sessions terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Prescreen (PUBLIC — no auth)
  # ===========================================================================
...
  /report/generateReport:
    put:
      tags: [Reports]
      summary: Generate an Excel pipeline report with filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilterRequest"
      responses:
        "200":
          description: Excel file stream
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /report/generatePdf:
    get:
      tags: [Reports]
      summary: Generate PDF feedback report for a specific interview
      parameters:
        - name: interviewerCalendarId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: PDF file stream or S3 URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Excel Downloads
  # ===========================================================================
  /downloadExcel/exportExcel:
    post:
      tags: [Reports]
      summary: Download main candidate pipeline Excel report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilterRequest"
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /downloadExcel/exportExcelPMO:
    post:
      tags: [Reports]
      summary: Download PMO demand report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilterRequest"
      responses:
        "200":
          description: PMO Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /downloadExcel/exportExcelL2AgingReport:
    post:
      tags: [Reports]
      summary: Download L2 aging report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilterRequest"
      responses:
        "200":
          description: L2 aging Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  # ===========================================================================
  # Lookup / Master Data
  # ===========================================================================
...
  /supplyScreen/getSupplyInfo:
    post:
      tags: [Supply]
      summary: Get recruitment supply matrix by skill and status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromDate:
                  type: string
                  format: date
                toDate:
                  type: string
                  format: date
                technologyId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Supply matrix data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Demand Screen
  # ===========================================================================
  /demandScreen/getDemandInfo:
    post:
      tags: [Demand]
      summary: Get hiring demand matrix bucketed by role start date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skillGroupId:
                  type: integer
                  format: int64
                technologyId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Demand matrix (pastdue, bill_one/two/three, strategic, pursuit buckets)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /demandUpload/upload:
    post:
      tags: [Demand]
      summary: Bulk upload demand data via Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Bench Screen
  # ===========================================================================
  /benchScreen/getBenchInfo:
    post:
      tags: [Bench]
      summary: Get bench employee matrix by skill and aging buckets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skillGroupId:
                  type: integer
                  format: int64
                technologyId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Bench matrix (6 aging buckets)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /benchUpload/upload:
    post:
      tags: [Bench]
      summary: Bulk upload bench data via Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"


  # ===========================================================================
  # Calendar / Scheduling (app-specific paths)
  # ===========================================================================
...
  /feedback/get/{interviewer_calendar_id}:
    get:
      tags: [Feedback]
      summary: Get feedback by interviewer calendar slot
      parameters:
        - name: interviewer_calendar_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Feedback data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /feedback/pdf/{interviewer_calendar_id}:
    get:
      tags: [Feedback]
      summary: Download feedback PDF
      parameters:
        - name: interviewer_calendar_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: PDF file
...